day35_haystack/myproject/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.6
django-haystack==3.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
